requirements.txt
Flask
openai
requests
psycopg2-binary